requirements.txt
Flask
pywin32
PyScanner